refactor(scripts): extract parseBackfillArgs to eliminate CPD duplication (ARCH-M1-12 / T-106)
Move shared parseArgs/parseBackfillLimit/guard boilerplate into a
shared parseBackfillArgs() helper in incident-country-backfill-report.ts
so both backfill scripts share a single implementation rather than
duplicating 23 lines that trigger SonarCloud CPD.
